// src/app/shared/models/detalle.ts
export interface Detalle{
    codigofacturadetalle: number;
    codigofactura: number;
    codigoproducto: string;
    actividadeconomica: string;
    codigoproductosin: number;
    descripcion: string;
    cantidad: number;
    unidadmedida: number;
    preciounitario: number;
    montodescuento: number;
    subtotal: number;
    cuenta: string;
    numeroSerie: string;
    numeroImei: string;
    codigoGrupo: string;
}